pyproject.toml
[project]
name = "bless-bot-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.13",
    "base58>=2.1.1",
    "better-proxy>=1.3.1",
    "bs4>=0.0.2",
    "curl-cffi>=0.11.4",
    "httpx[socks]>=0.28.1",
    "libp2p>=0.2.8",
    "loguru>=0.7.3",
    "pywasm>=2.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
    "tortoise-orm>=0.25.1",
    "typing-extensions>=4.14.0",
]
